feat(backend): Phase-2 DB foundation — schema + Alembic + RLS (FR-2.15)
Lands the persistence backbone needed for the rest of Phase 2 to slot
in as additive PRs. Three tables (`users`, `meetings`,
`transcript_segments`) with row-level security ENABLEd and FORCEd, an
async SQLAlchemy 2.0 engine wired through a FastAPI lifespan, and an
opt-in `AUTO_MIGRATE` entrypoint for compose-driven local dev.

Routes are unchanged — the WS still emits transcript events into the
void. Persistence wiring + auth land in the follow-up PRs.

Drive-by: pre-existing RUF046 in api/transcript.py, and a pre-existing
Dockerfile breakage where `.dockerignore` filtered out the README that
hatchling reads for the project's own metadata.

+++ backend/src/meeting_intelligence/api/transcript.py
@@ -132,7 +132,7 @@ def _percentile(values: list[int], q: float) -> int:
     if not values:
         return 0
     s = sorted(values)
-    rank = max(0, min(len(s) - 1, int(round(q * (len(s) - 1)))))
+    rank = max(0, min(len(s) - 1, round(q * (len(s) - 1))))
     return s[rank]
